Moving iree/base/main.h to iree/base/internal/.
